MeshAlgo : Split into several cpp / test files.
It remains a single header.
